Moved settings file to config dir.
